pvs:V1028 possible overflow. Cast operands

+++ src/preview.cpp
@@ -1008,7 +1008,7 @@ namespace {
 
     DataBuf makePnm(uint32_t width, uint32_t height, const DataBuf &rgb)
     {
-        const long expectedSize = static_cast<long>(width * height * 3);
+        const long expectedSize = static_cast<long>(width) * static_cast<long>(height) * 3L;
         if (rgb.size_ != expectedSize) {
 #ifndef SUPPRESS_WARNINGS
             EXV_WARNING << "Invalid size of preview data. Expected " << expectedSize << " bytes, got " << rgb.size_ << " bytes.\n";
